requirements.txt
pydantic-ai
python-dotenv
ddgs
crawl4ai
fastmcp
